// frontend/src/app/settings/models/settings.model.ts
export interface CompanySettings {
  id?: string;
  company_name?: string;
  company_email?: string;
  company_phone?: string;
  bank_account_name?: string;
  bank_name?: string;
  account_number?: string;
  sort_code?: string;
  iban?: string;
  user_id?: string;
  created_at?: string;
  updated_at?: string;
}